test: isolate PROJECT fixtures and verify locked-context inactivation
