+++ backend/tests/test_ci_test_lanes.py
@@ -73,6 +73,7 @@ def test_measured_hotspots_have_explicit_semantic_owners() -> None:
         "tests/projects/guide_compilation/test_repository_attempts.py",
         "tests/projects/guide_compilation/test_repository_persistence.py",
         "tests/projects/test_locked_policy_context.py",
+        "tests/projects/test_locked_policy_contract.py",
         "tests/projects/test_activation_readiness.py",
         "tests/projects/test_retired_submission_derivation_route.py",
         "tests/test_projects.py",
